rename azure ssh provider to ssh-bastion
