+++ src/plugins/azure/ssh-bastion.ts
